pyproject.toml
[project]
name = "spot-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"fastapi>=0.95.0",
	"uvicorn[standard]>=0.22.0",
	"azure-identity>=1.14.0",
	"azure-mgmt-compute>=29.0.0",
	"cachetools>=5.3.0",
	"azure-mgmt-subscription>=1.0.0",
	"pytest>=8.4.1",
	"pre-commit>=4.3.0",
]
